style: 에러 핸들링 파일 분리
