+++ .github/workflows/nodejs.yml
@@ -16,6 +16,6 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
       - run: yarn
-      - run: yarn test
+      - run: NODE_ENV=sample yarn test
       - run: yarn lint
