+++ src/test/resolver_handler/Session.test.ts
@@ -1,7 +1,8 @@
 import sinon from 'sinon';
 import * as sessionHandler from '../../api/graphql/resolver_handler/Session';
 import getDatabase from '../../database';
-import * as constants from '../../common/constants';
+import * as errorHandler from '../../common/error';
+
 
 const database = getDatabase(true);
 const db = database.getInstance();
@@ -130,7 +131,7 @@ describe('resolver_handler [Session]', () => {
       session_explainer: 'temp',
       document: 'temp',
     });
-    expect({ statusCode: '106', errorMessage: constants.CustomError.MESSAGE['106'] }).toEqual(result);
+    expect({ statusCode: '106', errorMessage: errorHandler.CustomError.MESSAGE['106'] }).toEqual(result);
   });
 
   it('is the case when user have permission <updateSession>', async () => {
@@ -142,7 +143,7 @@ describe('resolver_handler [Session]', () => {
   it('is the case when user do not have permission <updateSession>', async () => {
     const context = { request: { session: { passport: { user: { email: 'test2@gmail.com' } } } } };
     const result = await sessionHandler.updateSession({}, { user_email: 'test@gmail.com' }, context);
-    expect({ statusCode: '104', errorMessage: constants.CustomError.MESSAGE['104'] }).toEqual(result);
+    expect({ statusCode: '104', errorMessage: errorHandler.CustomError.MESSAGE['104'] }).toEqual(result);
   });
 
   it('is the case when user have permission <deleteSession>', async () => {
@@ -154,6 +155,6 @@ describe('resolver_handler [Session]', () => {
   it('is the case when user do not have permission <deleteSession>', async () => {
     const context = { request: { session: { passport: { user: { email: 'test2@gmail.com' } } } } };
     const result = await sessionHandler.deleteSession({ user_email: 'test@gmail.com' }, context);
-    expect({ statusCode: '104', errorMessage: constants.CustomError.MESSAGE['104'] }).toEqual(result);
+    expect({ statusCode: '104', errorMessage: errorHandler.CustomError.MESSAGE['104'] }).toEqual(result);
   });
 });
